requirements.txt
Flask
requests
validators
shodan
python-dotenv
